Added code bits for cell volume
Includes:
ImageJ macro for finding, tracking, and measuring embryos in the stack.
Python script, with mashup of a function to convert a series of 48-bit
color tiffs to a series of 16-bit tiffs (grayscale), and commented out
script to run that for our directories, and code to import and group the
data from the Image J script.

The python code should be split into two scripts. The Image J macro
should be updated to deal with the odd way that Image J handles the x-y
position and angle for Feret measurement (simples just to make it also
measure the x-y center of the selection, and use that to group embryos).